deprecate `archive` functions, incorporate simplified `sync`, FileArchive and changes from the recent spike
