--- test/data/test_recenttracks.json
@@ -1,1 +1,0 @@
-{"recenttracks":{"track":[{"artist":{"#text":"Amy Macdonald","mbid":"6aa40207-fec8-43a7-991d-b872a42def05"},"name":"Footballer's Wife","streamable":"0","mbid":"36c1833e-76e8-4606-928f-0489d3e641d6","album":{"#text":"This Is the Life","mbid":"4846826f-c71e-4172-9229-4e1ff7d3e033"},"url":"https://www.last.fm/music/Amy+Macdonald/_/Footballer%27s+Wife","image":[{"#text":"https://lastfm-img2.akamaized.net/i/u/34s/8c84a8b2432e4810a87a8640d72af5c0.png","size":"small"},{"#text":"https://lastfm-img2.akamaized.net/i/u/64s/8c84a8b2432e4810a87a8640d72af5c0.png","size":"medium"},{"#text":"https://lastfm-img2.akamaized.net/i/u/174s/8c84a8b2432e4810a87a8640d72af5c0.png","size":"large"},{"#text":"https://lastfm-img2.akamaized.net/i/u/300x300/8c84a8b2432e4810a87a8640d72af5c0.png","size":"extralarge"}],"date":{"uts":"1199120880","#text":"31 Dec 2007, 17:08"}}],"@attr":{"user":"lastfm_user","page":1,"perPage":1,"totalPages":12,"total":12}}}
