--- test/data/test_recenttracks_no_scrobble.json
@@ -1,1 +1,0 @@
-{"recenttracks":{"track":[],"@attr":{"user":"a_lastfm_user","page":1,"perPage":1,"totalPages":0,"total":0}}}
