--- test/data/test_user.json
@@ -1,1 +1,0 @@
-{"user":{"playlists":"0","playcount":1234,"name":"lastfm_user","subscriber":"0","url":"https:\/\/www.last.fm\/user\/lastfm_user","country":"United Kingdom","registered":{"unixtime":1472601600,"#text":1472601600},"type":"user","age":"0","bootstrap":"0","realname":"Test User"}}
