--- test/data/test_user2.json
@@ -1,1 +1,0 @@
-{"user":{"playlists":"0","playcount":1234,"name":"lastfm_user","subscriber":"0","url":"https:\/\/www.last.fm\/user\/lastfm_user","country":"United Kingdom","registered":{"unixtime":1483228800,"#text":1483228800},"type":"user","age":"0","bootstrap":"0","realname":"Test User"}}
